[ refactor ] Code Refactoring & Cleanup
