+++ ThinkerBell/app/build.gradle.kts
@@ -61,4 +61,5 @@ dependencies {
 
     implementation(libs.androidx.datastore.preferences)
 
+    implementation(libs.glide)
 }
