+++ ThinkerBell/app/src/main/java/com/neverland/thinkerbell/presentation/view/home/HomeNoticeFragment.kt
@@ -1,0 +1,23 @@
+package com.neverland.thinkerbell.presentation.view.home
+
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.neverland.thinkerbell.R
+import com.neverland.thinkerbell.databinding.FragmentHomeNoticeBinding
+import com.neverland.thinkerbell.domain.model.Notice
+import com.neverland.thinkerbell.presentation.view.home.adapter.HomeNoticeAdapter
+import com.neverland.thinkerbell.presentation.base.BaseFragment
+import com.neverland.thinkerbell.presentation.view.CustomDividerItemDecoration
+
+
+class HomeNoticeFragment(
+    private val notices: List<Notice>
+) : BaseFragment<FragmentHomeNoticeBinding>(R.layout.fragment_home_notice) {
+
+    override fun initView() {
+        binding.rvHomeNotice.layoutManager = LinearLayoutManager(context)
+        binding.rvHomeNotice.adapter = HomeNoticeAdapter(notices)
+
+        val customDividerItemDecoration = CustomDividerItemDecoration(requireContext())
+        binding.rvHomeNotice.addItemDecoration(customDividerItemDecoration)
+    }
+}
