+++ ThinkerBell/app/src/main/res/layout/activity_home.xml
@@ -5,16 +5,18 @@
     <data>
 
     </data>
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/primary2">
         <FrameLayout
-            android:id="@+id/fl_main"
+            android:id="@+id/fl_home"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toTopOf="@id/shadow_view"/>
+
         <!-- 그림자 효과를 위한 View -->
         <View
             android:id="@+id/shadow_view"
