+++ ThinkerBell/app/src/main/res/layout/item_calendar_day.xml
@@ -1,4 +1,5 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/ll_day"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
@@ -8,10 +9,10 @@
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/tv_day"
         android:layout_width="35dp"
-        android:layout_height="19.2dp"
+        android:layout_height="20dp"
         style="@style/Paragraph_Small"
-        android:text="9"
+        tools:text="9"
         android:gravity="center"
-        android:layout_marginLeft="8dp"
+        android:layout_marginStart="8dp"
         android:textColor="@color/content_secondary"/>
 </LinearLayout>
