+++ ThinkerBell/app/src/main/res/layout/item_calendar_month.xml
@@ -12,7 +12,7 @@
         android:layout_height="wrap_content"
         android:src="@drawable/ic_direaction_left_1"
         android:background="?attr/selectableItemBackgroundBorderless"
-        android:tint="@color/content_secondary"
+        app:tint="@color/content_secondary"
         android:contentDescription="Previous Month"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="@id/tv_month"
@@ -42,7 +42,7 @@
         android:layout_height="wrap_content"
         android:src="@drawable/ic_direaction_right_1"
         android:background="?attr/selectableItemBackgroundBorderless"
-        android:tint="@color/content_secondary"
+        app:tint="@color/content_secondary"
         android:contentDescription="Next Month"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="@id/tv_month"
